Location Tracking, loading viewsets
Viewset loading took a fair amount of time. I initially used the unitywebrequest class, but it took issue when downloading archives from the rest API (would claim authorization headers to be unnecessary when present, required when not). I continued trying to work with it for a bit too long, because it seemed to work with all other requests. Eventually, I switched to using the HttpWebRequest, which worked fine.

Location tracking was reasonably quick to get to this point. Functional in the editor when you connect an IOS device through unity remote. Currently, the service is enabled/disabled as needed (in the name of battery efficiency) which makes it pretty slow. Will fix this next push, once I think of something better.